+++ frameworks/core/interfaces/native/node/node_textpicker_modifier.cpp
@@ -78,7 +78,7 @@ void GetTextPickerSelectedIndex(ArkUINodeHandle node, ArkUI_Uint32* values, ArkU
         values[0] = selectedValue;
     } else {
         auto selectedValue = TextPickerModelNG::getTextPickerSelecteds(frameNode);
-        if (size != static_cast<int32_t>(selectedValue.size())) {
+        if (size != selectedValue.size()) {
             return;
         }
         for (int i = 0; i < size; i++) {
